requirements.txt
amqp==5.2.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.4.0
async-timeout==4.0.3
autopep8==2.0.2
backcall==0.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.5
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cssselect2==0.7.0
decorator==5.1.1
Django==4.2.13
django-cors-headers==4.0.0
django-debug==0.1.1
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-redis==5.4.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
dub==0.7.3
exceptiongroup==1.1.3
executing==2.0.0
fonttools==4.47.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.16.1
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
kombu==5.3.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
openapi-codec==1.3.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
# psycopg==3.1.8
psycopg>=3.1.8
# psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.8.0
Pygments==2.16.1
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.1
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tinycss2==1.2.1
tomli==2.0.1
traitlets==5.11.2
typing_extensions>=4.6.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.8
weasyprint==60.2
webencodings==0.5.1
zopfli==0.2.3
